+++ src/Resources/Users/index.ts
@@ -1,6 +1,7 @@
 import Client from '../../Client';
 import AbstractResource from '../AbstractResource';
 import Baskets from './Baskets';
+import Gifts from './Gifts';
 import Subscriptions from './Subscriptions';
 import Toybox from './Toybox';
 import Details from './Details';
@@ -9,6 +10,7 @@ import Addresses from './Addresses';
 export default class Users extends AbstractResource {
     public addresses: Addresses;
     public baskets: Baskets;
+    public gifts: Gifts;
     public details: Details;
     public subscriptions: Subscriptions;
     public toybox: Toybox;
@@ -17,8 +19,10 @@ export default class Users extends AbstractResource {
         super(api);
         this.api = api;
         this.initialise();
+
         this.addresses = new Addresses(api);
         this.baskets = new Baskets(api);
+        this.gifts = new Gifts(api);
         this.details = new Details(api);
         this.subscriptions = new Subscriptions(api);
         this.toybox = new Toybox(api);